pyproject.toml
[project]
name = "sitcon-marketing-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot[job-queue]>=20.0",
]
